str_net: fix loading of npz file under python3
That file is contains an array of objects (numpy arrays), so
pickling is used. Unfortunately that is much more problematic
than an .npz file with just normal numpy arrays.

While at it, assume that the spikes are stored under 'spikeTime'.
This actually makes things more flexible, because other items can
be added to the same .npz file without breaking this code.