Changed NamedDict to allow setting values via dict or attribute method.
It is now possible to override parameters created in NamedDicts by
doing myNamedDict[key] = newvalue OR myNamedDict.key = newvalue.
Both the attribute method or the dict key method sychronize values.
New keys/attributes can now also be created: myNamedDict[newkey] =
value and myNamedDict.newkey = value are both valid. Note that keys
should be limited to valid attribute names, that is, should be strings
beginning with a letter. Attributes beginning with an underscore are
also valid as attributes but NOT included in the dictionary (e.g.
__name__).

Prior to this change, only mutable types (e.g. dictionaries in
Condset NamedDict) could be changed, but not immutable types (like
float/int/strings, such as SpineParams.density).