Added flexible command line option for model param overrides.
Any parameter that is a string or number AND that is accessible
as attributes of parent model, e.g. model.MyParamDict.MyParamName
with dot notation can be specified for command line ovverride as:

--modelParamOverrides MyParamDict.MyParamName1:NewValue1 MyParamDict.MyParamName2:NewValue2

Note that param:value pairs must be colonseparated, and any number
of pairs can be given as space-separated.

Might be useful to add list/dictionary notation with brackets in
the future, e.g.: MyParamDict.MyParamList[1]:NewValue